Move Bitcoin Address and Uint256 json converter to namespace json
