+++ src/Ztm.Zcoin.NBitcoin/Json/UInt256Converter.cs
@@ -2,9 +2,9 @@ using System;
 using NBitcoin;
 using Newtonsoft.Json;
 
-namespace Ztm.Zcoin.NBitcoin
+namespace Ztm.Zcoin.NBitcoin.Json
 {
-    public sealed class UInt256JsonConverter : JsonConverter<uint256>
+    public sealed class UInt256Converter : JsonConverter<uint256>
     {
         public override uint256 ReadJson(JsonReader reader, Type objectType, uint256 existingValue, bool hasExistingValue, JsonSerializer serializer)
         {
